requirements.txt
Flask==2.3.4
Flask-Mail==0.9.1
Werkzeug==2.3.6
python-dotenv==1.0.0
